pyproject.toml
[project]
name = "investment-brief"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "feedparser==6.0.12",
    "finnhub-python==2.4.27",
    "python-dotenv==1.2.2",
    "rich==14.3.3",
    "schedule==1.2.2",
    "yfinance==1.2.0",
]
